requirements.txt
streamlit>=1.35.0
groq>=1.2.0
python-dotenv>=1.0.0